Split ProductIdentity tests to separate TestFixture
